requirements.txt
absl-py==0.11.0
appnope==0.1.0
attrs==19.3.0
audioread==2.1.8
av==8.0.2
backcall==0.1.0
cffi==1.14.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
importlib-metadata==1.6.0
inflect==4.1.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
jiwer==2.1.0
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
numba==0.48.0
numpy==1.18.4
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pycparser==2.20
pydub==0.24.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sounddevice==0.3.15
SoundFile==0.10.3.post1
speechpy==2.4
sphfile==1.0.3
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.8.0.dev1
torchaudio==0.4.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
Unidecode==1.1.1
wcwidth==0.2.3
webencodings==0.5.1
youtube-dl==2020.11.18
gradio==2.0.1